+++ src/components/application_manager/src/policies/policy_handler.cc
@@ -1941,7 +1941,7 @@ void PolicyHandler::GetAppServiceParameters(
     const std::string& policy_app_id,
     std::string& service_name,
     std::string& service_type,
-    std::vector<uint64_t>& handled_rpcs) const {
+    std::vector<uint32_t>& handled_rpcs) const {
   POLICY_LIB_CHECK_VOID();
   policy_manager_->GetAppServiceParameters(
       policy_app_id, service_name, service_type, handled_rpcs);
